Crate locale_config

Global locale instances and system inspection.

This is an auxiliary crate for i18n solutions that:

Holds the appropriate default instances of locale.

Inspects the system for the initial values.

You don’t want to use it directly, but instead use an internationalisation crate like locale.

This crate is separate and intentionally minimal so that multiple i18n crates or multiple versions of one that get into the application still share the current locale setting.

Structs

LanguageRange

Language and culture identifier.

Locale

Locale configuration.

Tags

Iterator over LanguageRanges for all categories in a Locale

TagsFor

Iterator over LanguageRanges for specific category in a Locale

Enums

Error

Errors that may be returned by locale_config.