Struct locale_config::Locale

pub struct Locale { /* fields omitted */ }

Locale configuration.

Users may accept several languages in some order of preference and may want to use rules from different culture for some particular aspect of the program behaviour, and operating systems allow them to specify this (to various extent).

The Locale objects represent the user configuration. They contain:

The primary LanguageRange.

Optional category-specific overrides.

Optional fallbacks in case data (usually translations) for the primary language are not available.

The set of categories is open-ended. The locale crate uses five well-known categories messages, numeric, time, collate and monetary, but some systems define additional ones (GNU Linux has additionally paper, name, address, telephone and measurement) and these are provided in the user default Locale and other libraries can use them.

Locale is represented by a ,-separated sequence of tags in LanguageRange syntax, where all except the first one may be preceded by category name and = sign.

The first tag indicates the default locale, the tags prefixed by category names indicate overrides for those categories and the remaining tags indicate fallbacks.

Note that a syntactically valid value of HTTP Accept-Language header is a valid Locale. Not the other way around though due to the presence of category selectors.

Methods

impl Locale

pub fn user_default() -> Locale

Obtain the user default locale.

This is the locale indicated by operating environment.

pub fn global_default() -> Locale

Obtain the global default locale.

The global default for current() locale. Defaults to user_default().

pub fn set_global_default(lb: Locale)

Change the global default locale.

Setting this overrides the default for new threads and threads that didn't do any locale-aware operation yet.

pub fn current() -> Locale

Obtain the current locale of current thread.

Defaults to global_default() on first use in each thread.

pub fn set_current(lb: Locale)

Change the current locale of current thread.

pub fn new(s: &str) -> Result<Locale, Error>

Construct locale from the string representation.

Locale is represented by a ,-separated sequence of tags in LanguageRange syntax, where all except the first one may be preceded by category name and = sign.

The first tag indicates the default locale, the tags prefixed by category names indicate overrides for those categories and the remaining tags indicate fallbacks.

pub fn invariant() -> Locale

Construct invariant locale.

Invariant locale is represented simply with empty string.

pub fn add(&mut self, tag: &LanguageRange)

Append fallback language tag.

Adds fallback to the end of the list.

pub fn add_category(&mut self, category: &str, tag: &LanguageRange)

Append category override.

Appending new override for a category that already has one will not replace the existing override. This might change in future.

Important traits for Tags<'a>
impl<'a> Iterator for Tags<'a> type Item = (Option<&'a str>, LanguageRange<'a>);

pub fn tags<'a>(&'a self) -> Tags<'a>

Iterate over LanguageRanges in this Locale.

Returns tuples of optional category (as string) and corresponding LanguageRange. All tags in the list are returned, in order of preference.

The iterator is guaranteed to return at least one value.

Important traits for TagsFor<'a, 'c>
impl<'a, 'c> Iterator for TagsFor<'a, 'c> type Item = LanguageRange<'a>;

pub fn tags_for<'a, 'c>(&'a self, category: &'c str) -> TagsFor<'a, 'c>

Iterate over LanguageRanges in this Locale applicable to given category.

Returns LanguageRanges in the Locale that are applicable to provided category. The tags are returned in order of preference, which means the category-specific ones first and then the generic ones.

The iterator is guaranteed to return at least one value.